requirements.txt

gradio
gTTS
openai
python-dotenv
